@@ -3,6 +3,7 @@ import Foundation
 import Backtrace
 import CBacktrace
 #endif
+import NIOConcurrencyHelpers
 
 extension Optional where Wrapped == StackTrace {
     public static func capture(skip: Int = 0) -> Self {
@@ -10,8 +11,20 @@ extension Optional where Wrapped == StackTrace {
     }
 }
 
-public struct StackTrace {
-    public static var isCaptureEnabled: Bool = true
+public struct StackTrace: Sendable {
+    public static var isCaptureEnabled: Bool {
+        get {
+            _isCaptureEnabled.withLockedValue {
+                $0
+            }
+        }
+        set {
+            _isCaptureEnabled.withLockedValue {
+                $0 = newValue
+            }
+        }
+    }
+    private static let _isCaptureEnabled: NIOLockedValueBox<Bool> = .init(true)
 
     public static func capture(skip: Int = 0) -> Self? {
         guard Self.isCaptureEnabled else {
@@ -58,7 +71,7 @@ public struct StackTrace {
         #endif
     }
 
-    public struct Frame {
+    public struct Frame: Sendable {
         public var file: String
         public var function: String
     }
@@ -72,7 +85,7 @@ public struct StackTrace {
         }
     }
 
-    struct RawFrame {
+    struct RawFrame: Sendable {
         var file: String
         var mangledFunction: String
     }
