@@ -1,5 +1,5 @@
 import NIOCore
 
-public protocol Responder {
+public protocol Responder: Sendable {
     func respond(to request: Request) -> EventLoopFuture<Response>
 }
