@@ -1,6 +1,6 @@
 import Foundation
 
-public protocol PasswordHasher {
+public protocol PasswordHasher: Sendable {
     func hash<Password>(_ password: Password) throws -> [UInt8]
         where Password: DataProtocol
 
