@@ -1,6 +1,7 @@
 import Foundation
 import NIOPosix
 import NIOCore
+import NIOConcurrencyHelpers
 
 /// An internal helper that formats cookie dates as RFC1123
 private final class RFC1123 {
@@ -58,7 +59,7 @@ extension Date {
 }
 
 /// Performant method for generating RFC1123 date headers.
-internal final class RFC1123DateCache {
+internal final class RFC1123DateCache: Sendable {
     static func eventLoop(_ eventLoop: EventLoop) -> RFC1123DateCache {
         assert(eventLoop.inEventLoop)
         
@@ -79,20 +80,17 @@ internal final class RFC1123DateCache {
     /// Thread-specific RFC1123
     private static let thread: ThreadSpecificVariable<RFC1123DateCache> = .init()
     
-    /// Currently cached time components.
-    private var cachedTimeComponents: (key: time_t, components: tm)?
-    
-    /// Currently cached timestamp.
-    private var timestamp: String
+    /// Currently cached time components and timestamp
+    private let cachedTimestampAndComponents: NIOLockedValueBox<((key: time_t, components: tm)?, String)>
     
     /// Creates a new `RFC1123DateCache`.
     private init() {
-        self.timestamp = ""
+        self.cachedTimestampAndComponents = .init((nil, ""))
         self.updateTimestamp()
     }
     
     func currentTimestamp() -> String {
-        return self.timestamp
+        return self.cachedTimestampAndComponents.withLockedValue { $0.1 }
     }
     
     /// Updates the current RFC 1123 date string.
@@ -104,50 +102,53 @@ internal final class RFC1123DateCache {
         // this key is a unique id for each day
         let key = date / secondsInDay
         
-        // get time components
-        let dateComponents: tm
-        if let cached = self.cachedTimeComponents, cached.key == key {
-            dateComponents = cached.components
-        } else {
-            var tc = tm.init()
-            gmtime_r(&date, &tc)
-            dateComponents = tc
-            self.cachedTimeComponents = (key: key, components: tc)
+        self.cachedTimestampAndComponents.withLockedValue { cachedValues in
+            // get time components
+            let dateComponents: tm
+            
+            if let cachedTimeComponents = cachedValues.0, cachedTimeComponents.key == key {
+                dateComponents = cachedTimeComponents.components
+            } else {
+                var tc = tm.init()
+                gmtime_r(&date, &tc)
+                dateComponents = tc
+                cachedValues.0 = (key: key, components: tc)
+            }
+            
+            // parse components
+            let year: Int = numericCast(dateComponents.tm_year) + 1900 // years since 1900
+            let month: Int = numericCast(dateComponents.tm_mon) // months since January [0-11]
+            let monthDay: Int = numericCast(dateComponents.tm_mday) // day of the month [1-31]
+            let weekDay: Int = numericCast(dateComponents.tm_wday) // days since Sunday [0-6]
+            
+            // get basic time info
+            let t: Int = date % secondsInDay
+            let hours: Int = numericCast(t / 3600)
+            let minutes: Int = numericCast((t / 60) % 60)
+            let seconds: Int = numericCast(t % 60)
+            
+            // generate the RFC 1123 formatted string
+            var rfc1123 = ""
+            rfc1123.reserveCapacity(30)
+            rfc1123.append(dayNames[weekDay])
+            rfc1123.append(", ")
+            rfc1123.append(stringNumbers[monthDay])
+            rfc1123.append(" ")
+            rfc1123.append(monthNames[month])
+            rfc1123.append(" ")
+            rfc1123.append(stringNumbers[year / 100])
+            rfc1123.append(stringNumbers[year % 100])
+            rfc1123.append(" ")
+            rfc1123.append(stringNumbers[hours])
+            rfc1123.append(":")
+            rfc1123.append(stringNumbers[minutes])
+            rfc1123.append(":")
+            rfc1123.append(stringNumbers[seconds])
+            rfc1123.append(" GMT")
+            
+            // cache the new timestamp
+            cachedValues.1 = rfc1123
         }
-        
-        // parse components
-        let year: Int = numericCast(dateComponents.tm_year) + 1900 // years since 1900
-        let month: Int = numericCast(dateComponents.tm_mon) // months since January [0-11]
-        let monthDay: Int = numericCast(dateComponents.tm_mday) // day of the month [1-31]
-        let weekDay: Int = numericCast(dateComponents.tm_wday) // days since Sunday [0-6]
-        
-        // get basic time info
-        let t: Int = date % secondsInDay
-        let hours: Int = numericCast(t / 3600)
-        let minutes: Int = numericCast((t / 60) % 60)
-        let seconds: Int = numericCast(t % 60)
-        
-        // generate the RFC 1123 formatted string
-        var rfc1123 = ""
-        rfc1123.reserveCapacity(30)
-        rfc1123.append(dayNames[weekDay])
-        rfc1123.append(", ")
-        rfc1123.append(stringNumbers[monthDay])
-        rfc1123.append(" ")
-        rfc1123.append(monthNames[month])
-        rfc1123.append(" ")
-        rfc1123.append(stringNumbers[year / 100])
-        rfc1123.append(stringNumbers[year % 100])
-        rfc1123.append(" ")
-        rfc1123.append(stringNumbers[hours])
-        rfc1123.append(":")
-        rfc1123.append(stringNumbers[minutes])
-        rfc1123.append(":")
-        rfc1123.append(stringNumbers[seconds])
-        rfc1123.append(" GMT")
-        
-        // cache the new timestamp
-        self.timestamp = rfc1123
     }
 }
 
