@@ -1,6 +1,6 @@
 import NIOHTTP1
 
-public struct ValidationsResult {
+public struct ValidationsResult: Sendable {
     public let results: [ValidationResult]
     
     public var error: ValidationsError? {
@@ -19,7 +19,7 @@ public struct ValidationsResult {
     }
 }
 
-public struct ValidationsError: Error {
+public struct ValidationsError: Sendable, Error {
     public let failures: [ValidationResult]
 }
 
