@@ -1,6 +1,6 @@
-public struct Validator<T: Decodable> {
-    public let validate: (_ data: T) -> ValidatorResult
-    public init(validate: @escaping (_ data: T) -> ValidatorResult) {
+public struct Validator<T: Decodable & Sendable>: Sendable {
+    public let validate: @Sendable (_ data: T) -> ValidatorResult
+    public init(validate: @Sendable @escaping (_ data: T) -> ValidatorResult) {
         self.validate = validate
     }
 }
