@@ -1,3 +1,5 @@
+import NIOCore
+
 extension Validator where T: Comparable & Strideable {
     /// Validates that the data is within the supplied `Range`.
     public static func range(_ range: Swift.Range<T>) -> Validator<T> {
@@ -10,12 +12,12 @@ extension Validator where T: Comparable {
     public static func range(_ range: ClosedRange<T>) -> Validator<T> {
         .range(min: range.lowerBound, max: range.upperBound)
     }
-
+    
     /// Validates that the data is less than or equal to the supplied upper bound using `PartialRangeThrough`.
     public static func range(_ range: PartialRangeThrough<T>) -> Validator<T> {
         .range(min: nil, max: range.upperBound)
     }
-
+    
     /// Validates that the data is greater than or equal the supplied lower bound using `PartialRangeFrom`.
     public static func range(_ range: PartialRangeFrom<T>) -> Validator<T> {
         .range(min: range.lowerBound, max: nil)
@@ -38,10 +40,11 @@ extension Validator {
         min: U?, max: U?, _ keyPath: KeyPath<T, U>,
         _ suffix: String? = nil
     ) -> Validator<T>
-        where U: Comparable
-    {   
-        .init { data in 
-            if let result = try? RangeResult.init(min: min, max: max, value: data[keyPath: keyPath]) {
+    where U: Comparable & Sendable
+    {
+        let sendableKeyPath = UnsafeMutableTransferBox(keyPath)
+        return .init { data in
+            if let result = try? RangeResult.init(min: min, max: max, value: data[keyPath: sendableKeyPath.wrappedValue]) {
                 return ValidatorResults.Range(
                     result: result,
                     suffix: suffix
@@ -53,9 +56,26 @@ extension Validator {
     }
 }
 
+// TODO: Remove when keypaths are `Sendable`
+/// ``UnsafeMutableTransferBox`` can be used to make non-`Sendable` values `Sendable` and mutable.
+/// It can be used to capture local mutable values in a `@Sendable` closure and mutate them from within the closure.
+/// As the name implies, the usage of this is unsafe because it disables the sendable checking of the compiler and does not add any synchronisation.
+@usableFromInline
+final class UnsafeMutableTransferBox<Wrapped> {
+    @usableFromInline
+    var wrappedValue: Wrapped
+    
+    @inlinable
+    init(_ wrappedValue: Wrapped) {
+        self.wrappedValue = wrappedValue
+    }
+}
+
+extension UnsafeMutableTransferBox: @unchecked Sendable {}
+
 extension ValidatorResults {
     /// `ValidatorResult` of a validator that validates whether the input is within a supplied range.
-    public struct Range<T> where T: Comparable {
+    public struct Range<T> where T: Comparable & Sendable {
         /// The position of the data relative to the range.
         public let result: RangeResult<T>
         
