@@ -13,7 +13,7 @@ extension Authenticatable {
     ///
     /// - parameters:
     ///    - makePath: The closure that returns the redirect path based on the given `Request` object
-    public static func redirectMiddleware(makePath: @escaping (Request) -> String) -> Middleware {
+    @preconcurrency public static func redirectMiddleware(makePath: @Sendable @escaping (Request) -> String) -> Middleware {
         RedirectMiddleware<Self>(Self.self, makePath: makePath)
     }
 }
@@ -22,9 +22,9 @@ extension Authenticatable {
 private final class RedirectMiddleware<A>: Middleware
     where A: Authenticatable
 {
-    let makePath: (Request) -> String
+    let makePath: @Sendable (Request) -> String
     
-    init(_ authenticatableType: A.Type = A.self, makePath: @escaping (Request) -> String) {
+    @preconcurrency init(_ authenticatableType: A.Type = A.self, makePath: @Sendable @escaping (Request) -> String) {
         self.makePath = makePath
     }
 
