@@ -3,7 +3,7 @@ import NIOCore
 /// A basic, async closure-based `Responder`.
 public struct AsyncBasicResponder: AsyncResponder {
     /// The stored responder closure.
-    private let closure: (Request) async throws -> Response
+    private let closure: @Sendable (Request) async throws -> Response
 
     /// Create a new `BasicResponder`.
     ///
@@ -15,7 +15,7 @@ public struct AsyncBasicResponder: AsyncResponder {
     /// - parameters:
     ///     - closure: Responder closure.
     public init(
-        closure: @escaping (Request) async throws -> Response
+        closure: @Sendable @escaping (Request) async throws -> Response
     ) {
         self.closure = closure
     }
