@@ -3,7 +3,7 @@ import Foundation
 
 extension AsyncPasswordHasher {
     public func hash<Password>(_ password: Password) async throws -> [UInt8]
-        where Password: DataProtocol
+        where Password: DataProtocol & Sendable
     {
         try await self.hash(password).get()
     }
@@ -12,7 +12,7 @@ extension AsyncPasswordHasher {
         _ password: Password,
         created digest: Digest
     ) async throws -> Bool
-        where Password: DataProtocol, Digest: DataProtocol
+        where Password: DataProtocol & Sendable, Digest: DataProtocol & Sendable
     {
         try await self.verify(password, created: digest).get()
     }
