@@ -7,10 +7,11 @@ import Foundation
 extension Request {
 
     /// Upgrades an existing request to a websocket connection
+    @preconcurrency
     public func webSocket(
         maxFrameSize: WebSocketMaxFrameSize = .`default`,
-        shouldUpgrade: @escaping ((Request) async throws -> HTTPHeaders?) = { _ in [:] },
-        onUpgrade: @escaping (Request, WebSocket) async -> ()
+        shouldUpgrade: @escaping (@Sendable (Request) async throws -> HTTPHeaders?) = { _ in [:] },
+        onUpgrade: @Sendable @escaping (Request, WebSocket) async -> ()
     ) -> Response {
         webSocket(
             maxFrameSize: maxFrameSize,
@@ -41,12 +42,13 @@ extension RoutesBuilder {
     ///       See `NIOWebSocketServerUpgrader`.
     ///   - onUpgrade: Closure to apply after web socket is upgraded successfully.
     /// - returns: `Route` instance for newly created web socket endpoint
+    @preconcurrency
     @discardableResult
     public func webSocket(
         _ path: PathComponent...,
         maxFrameSize: WebSocketMaxFrameSize = .`default`,
-        shouldUpgrade: @escaping ((Request) async throws -> HTTPHeaders?) = { _ in [:] },
-        onUpgrade: @escaping (Request, WebSocket) async -> ()
+        shouldUpgrade: @escaping (@Sendable (Request) async throws -> HTTPHeaders?) = { _ in [:] },
+        onUpgrade: @Sendable @escaping (Request, WebSocket) async -> ()
     ) -> Route {
         return self.webSocket(path, maxFrameSize: maxFrameSize, shouldUpgrade: shouldUpgrade, onUpgrade: onUpgrade)
     }
@@ -60,12 +62,13 @@ extension RoutesBuilder {
     ///       See `NIOWebSocketServerUpgrader`.
     ///   - onUpgrade: Closure to apply after web socket is upgraded successfully.
     /// - returns: `Route` instance for newly created web socket endpoint
+    @preconcurrency
     @discardableResult
     public func webSocket(
         _ path: [PathComponent],
         maxFrameSize: WebSocketMaxFrameSize = .`default`,
-        shouldUpgrade: @escaping ((Request) async throws -> HTTPHeaders?) = { _ in [:] },
-        onUpgrade: @escaping (Request, WebSocket) async -> ()
+        shouldUpgrade: @escaping (@Sendable (Request) async throws -> HTTPHeaders?) = { _ in [:] },
+        onUpgrade: @Sendable @escaping (Request, WebSocket) async -> ()
     ) -> Route {
         return self.on(.GET, path) { request -> Response in
             return request.webSocket(maxFrameSize: maxFrameSize, shouldUpgrade: shouldUpgrade, onUpgrade: onUpgrade)
@@ -74,12 +77,13 @@ extension RoutesBuilder {
 }
 
 extension WebSocket {
+    @preconcurrency
     public static func connect(
         to url: String,
         headers: HTTPHeaders = [:],
         configuration: WebSocketClient.Configuration = .init(),
         on eventLoopGroup: EventLoopGroup,
-        onUpgrade: @escaping (WebSocket) -> ()
+        onUpgrade: @Sendable @escaping (WebSocket) -> ()
     ) async throws {
         guard let url = URL(string: url) else {
             throw WebSocketClient.Error.invalidURL
@@ -93,12 +97,13 @@ extension WebSocket {
         )
     }
 
+    @preconcurrency
     public static func connect(
         to url: URL,
         headers: HTTPHeaders = [:],
         configuration: WebSocketClient.Configuration = .init(),
         on eventLoopGroup: EventLoopGroup,
-        onUpgrade: @escaping (WebSocket) -> ()
+        onUpgrade: @Sendable @escaping (WebSocket) -> ()
     ) async throws  {
         let scheme = url.scheme ?? "ws"
         return try await self.connect(
@@ -113,6 +118,7 @@ extension WebSocket {
         )
     }
 
+    @preconcurrency
     public static func connect(
         scheme: String = "ws",
         host: String,
@@ -121,7 +127,7 @@ extension WebSocket {
         headers: HTTPHeaders = [:],
         configuration: WebSocketClient.Configuration = .init(),
         on eventLoopGroup: EventLoopGroup,
-        onUpgrade: @escaping (WebSocket) -> ()
+        onUpgrade: @Sendable @escaping (WebSocket) -> ()
     ) async throws  {
         return try await WebSocketClient(
             eventLoopGroupProvider: .shared(eventLoopGroup),
