@@ -1,5 +1,6 @@
 import Foundation
 import MultipartKit
+import NIOConcurrencyHelpers
 
 /// Configures which ``Encoder``s and ``Decoder``s to use when interacting with data in HTTP messages.
 ///
@@ -16,7 +17,17 @@ import MultipartKit
 /// Most often, these configured coders are used to encode and decode types conforming to ``Content``.
 /// See the ``Content`` protocol for more information.
 public struct ContentConfiguration {
-    public static var global: ContentConfiguration = .default()
+    public static var global: ContentConfiguration {
+        get {
+            _global.withLockedValue { $0 }
+        }
+        set {
+            _global.withLockedValue {
+                $0 = newValue
+            }
+        }
+    }
+    private static let _global: NIOLockedValueBox<ContentConfiguration> = .init(.default())
     
     /// Creates a ``ContentConfiguration`` containing all of Vapor's default coders.
     public static func `default`() -> ContentConfiguration {
