@@ -1,10 +1,19 @@
 import Vapor
 import XCTest
+import NIOConcurrencyHelpers
 
 /**
  We recommend configuring this in your test’s `class func setUp()`
  */
-public var app: (() throws -> Application)!
+public var app: (@Sendable () throws -> Application)! {
+    get {
+        appBox.withLockedValue({ $0 })
+    }
+    set {
+        appBox.withLockedValue { $0 = newValue }
+    }
+}
+private let appBox: NIOLockedValueBox<(@Sendable () throws -> Application)?> = .init(nil)
 
 open class XCTVaporTests: XCTestCase {
     open var app: Application!
